[feat] 내가 저장한 이벤트 조회 및 응답 dto (#64)
