+++ src/main/java/hyangyu/server/dto/myPage/MyDisplayDto.java
@@ -10,6 +10,6 @@ import java.util.List;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class MyEventDto {
+public class MyDisplayDto {
     private List<EventDto> displays;
 }
