+++ src/main/java/hyangyu/server/dto/myPage/MyDisplayResponseDto.java
@@ -1,2 +1,11 @@
-package hyangyu.server.dto.myPage;public class MyDisplayResponseDto {
+package hyangyu.server.dto.myPage;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class MyDisplayResponseDto {
+    int status;
+    MyDisplayDto data;
 }
