+++ src/main/java/hyangyu/server/dto/myPage/MyFairDto.java
@@ -1,2 +1,15 @@
-package hyangyu.server.dto.myPage;public class MyFairDto {
+package hyangyu.server.dto.myPage;
+
+import hyangyu.server.dto.EventDto;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class MyFairDto {
+    private List<EventDto> fairs;
 }
